+++ src/main/java/org/bluepigeon/admin/controller/LoginController.java
@@ -100,83 +100,5 @@ public class LoginController {
 		}
 		
 	}
-	@POST
-	@Path("/builder")
-	@Produces(MediaType.APPLICATION_JSON)
-	public ResponseMessage validateBuilder(@FormParam("email") String email,@FormParam("password") String password)//, @FormParam("cityid") int cityid)
-//	public Registration getJson(@Context HttpServletRequest req)
-	{
-		ResponseMessage registration = new ResponseMessage();
-		System.out.println("email in controller : "+email);
-		System.out.println("password in controller : "+password);
-		
-		ubuilder = new Builder();
-		ubuilder.setEmail(email);
-		ubuilder.setPassword(password);
-		
-		serviceLoginValidate = new ServiceLoginValidate();
-		session = req.getSession();
-		
-		registration =  serviceLoginValidate.isValidBuilder(ubuilder);
-		
-		//if(registration.getStatus()!=0 || registration.getErrorMessage().getErrorState() == 8)
-		//{
-			session = req.getSession();
-			session.setAttribute("ubname", registration.getData());
-			//session.setAttribute("cityid", cityid);
-			System.out.println("Hi from controller");
-		//}	
-			return registration;	
-	}
-
-	@POST
-	@Path("/activate/builder")
-	public int activateBuilderAccount(@FormParam("id") int id,@FormParam("password") String password)
-	{
-		Builder registration = new Builder();
-		ubuilder = new Builder();
-		ubuilder.setId(id);
-		ubuilder.setPassword(password);
-		System.out.println("in controller : "+password);
-		serviceLoginValidate = new ServiceLoginValidate();
-		int validate = serviceLoginValidate.isActivateBuilderAccount(ubuilder);
-		return validate;
-		
-	}
 	
-	@GET
-	@Path("/logoutbuilder")
-	public Response logoutBuilder()
-	{
-		java.net.URI location1 = null;
-		try
-		{
-		location1 = new java.net.URI("../builder/index.jsp");
-		session = req.getSession(false);
-		if(session != null)
-		{
-			session.invalidate();
-			return Response.temporaryRedirect(location1).build();
-		}
-		else
-		{
-			return Response.temporaryRedirect(location1).build();
-		}
-		}
-		catch(Exception e)
-		{
-			e.printStackTrace();
-			return Response.temporaryRedirect(location1).build();
-			
-		}
-	}
-	@POST
-	@Path("/builder/password")
-	@Produces(MediaType.APPLICATION_JSON)
-	public ResponseMessage updateBuilderPassword(@FormParam("oldpassword") String oldPassword,@FormParam("password") String password){
-		System.out.println("Input password :: "+oldPassword);
-		BuilderDetailsDAO builderDetailsDAO = new BuilderDetailsDAO();
-		return builderDetailsDAO.updateBuilderPassword(oldPassword,password);
-	}
-
 }
