+++ src/main/java/org/bluepigeon/admin/dao/LoginValidateDAO.java
@@ -11,7 +11,4 @@ public interface LoginValidateDAO {
 //	public List<Registration> getAdminUserByEmailAndPassword(String eail, String password);
 	public  boolean isValidEmailId(String email);
 	public boolean isValidPassword(String password);
-	public Builder getBuilderByEmail(String email);
-	public boolean isValidBuilderEmailId(String email);
-	public boolean isValidBuilderPassword(String password);
 }
