+++ src/main/java/org/bluepigeon/admin/service/ServiceLoginValidate.java
@@ -114,108 +114,4 @@ public class ServiceLoginValidate implements ServiceLoginDAO {
 		LoginValidationImp loginValidationImp = new LoginValidationImp();	
 		return loginValidationImp.activateAccount(registration);
 	}
-	public ResponseMessage isValidBuilder(Builder registration) {
-		ResponseMessage response = new ResponseMessage();
-		try
-		{
-		if(registration.getEmail().trim().length()<=0 && registration.getPassword().trim().length()<=0)
-		{
-			response.setStatus(2);
-			response.setMessage("Please Enter your email id and password");
-			
-			return response;
-		}
-		else if(registration.getEmail().trim().length()<=0)
-		{
-			response.setStatus(2);
-			response.setMessage("Please enter your email id");
-			return response;
-		}
-		else if(registration.getPassword().trim().length()<=0)
-		{
-			response.setStatus(2);
-			response.setMessage("Please enter your password");
-			return response;
-		}
-		else 
-		{
-			LoginValidationImp loginValidationImp = new LoginValidationImp();
-			if(loginValidationImp.isValidBuilderEmailId(registration.getEmail()))
-			{
-				Builder emailList = loginValidationImp.getBuilderByEmail(registration.getEmail());
-				if(emailList != null)
-				{
-					ubuilder = new Builder();
-					ubuilder = emailList;
-					System.out.println("Builder Password :::: "+registration.getPassword());
-					System.out.println("DB Builder Password :::: "+ubuilder.getPassword());
-					if(ubuilder.getPassword().equals(registration.getPassword()))
-					{
-						System.out.println(":::::::::::::::Builder Builder Password ::::::::::");
-						System.out.println("DB PASS:"+ubuilder.getPassword());
-						System.out.println("Input PASS: "+registration.getPassword());
-						if(ubuilder.getLoginStatus()==1)
-						{
-							response.setStatus(1);
-							response.setMessage("loggedIn");
-							response.setData(ubuilder);
-							return response;
-						}
-						else if(ubuilder.getLoginStatus()==0)
-						{
-							response.setStatus(0);
-							response.setMessage("loggedIn");
-							response.setData(ubuilder);
-							return response;
-						}
-					}
-					else
-					{
-						ubuilder = new Builder();
-						ubuilder.setLoginStatus(2);
-						response.setStatus(2);
-						response.setMessage("Invalid password");
-						response.setData(ubuilder);
-						return response;
-					}
-				} else {
-					ubuilder = new Builder();
-					ubuilder.setLoginStatus(2);
-					response.setStatus(2);
-					response.setMessage("Please enter your valid email id");
-					response.setData(ubuilder);
-					return response;
-				}
-			} else if (loginValidationImp.isValidBuilderPassword(registration.getPassword())) {
-				ubuilder = new Builder();
-				ubuilder.setLoginStatus(2);
-				response.setStatus(2);
-				response.setMessage("Please enter your valid email id");
-				response.setData(ubuilder);
-				return response;
-			} else {
-			
-				ubuilder = new Builder();
-				ubuilder.setLoginStatus(2);
-				response.setStatus(2);
-				response.setMessage("Invalid user id and password");
-				response.setData(ubuilder);
-				return response;
-			}
-		
-		}
-		}
-		catch(NullPointerException e)
-		{
-			System.out.println(e);
-			e.printStackTrace();
-		}
-		return response;
-	}
-	public int isActivateBuilderAccount(Builder registration)
-	{
-		LoginValidationImp loginValidationImp = new LoginValidationImp();	
-		return loginValidationImp.activateBuilderAccount(registration);
-	}
-	
 }
