+++ src/main/resources/hibernate.cfg.xml
@@ -11,8 +11,8 @@
 		<property name="hibernate.connection.driver_class">com.mysql.jdbc.Driver</property>
 		<property name="hibernate.connection.username">root</property>
 <!-- 		<property name="hibernate.connection.password">olocafe@2017</property> -->
- 		<property name="hibernate.connection.password">Freight@123#</property> 
-<!-- 		<property name="hibernate.connection.password">root</property>   -->
+<!--  		<property name="hibernate.connection.password">Freight@123#</property>  -->
+		<property name="hibernate.connection.password">root</property>  
 		<property name="hibernate.connection.url">jdbc:mysql://localhost:3306/blue_pigeon</property>
 		<property name="hibernate.dialect">org.hibernate.dialect.MySQLDialect</property>
 		<property name="hibernate.cache.use_second_level_cache">false</property>
