+++ src/main/webapp/builder/agreement/new.jsp
@@ -10,15 +10,15 @@
 <%@page import="java.util.List"%>
 <%
 	session = request.getSession(false);
-	Builder builder = new Builder();
+	BuilderEmployee builder = new BuilderEmployee();
 	List<ProjectData> project_list = null;
 	int builder_id = 0;
 	if(session!=null)
 	{
 		if(session.getAttribute("ubname") != null)
 		{
-			builder  = (Builder)session.getAttribute("ubname");
-			builder_id = builder.getId();
+			builder  = (BuilderEmployee)session.getAttribute("ubname");
+			builder_id = builder.getBuilder().getId();
 		}
 		if(builder_id > 0){
 			project_list = new ProjectDAO().getActiveProjectsByBuilderId(builder_id);
