+++ src/main/webapp/builder/buyer/list.jsp
@@ -10,14 +10,14 @@
 <%@page import="java.util.List"%>
 <%
 	session = request.getSession(false);
-	Builder builder = new Builder();
+	BuilderEmployee builder = new BuilderEmployee();
 	int builder_id = 0;
 	if(session!=null)
 	{
 		if(session.getAttribute("ubname") != null)
 		{
-			builder  = (Builder)session.getAttribute("ubname");
-			builder_id = builder.getId();
+			builder  = (BuilderEmployee)session.getAttribute("ubname");
+			builder_id = builder.getBuilder().getId();
 		}
    }
 	List<Buyer> buyerList = new BuyerDAO().getAllBuyerByBuilderId(builder_id);
