+++ src/main/webapp/builder/buyer/new.jsp
@@ -8,15 +8,15 @@
 <%@page import="java.util.List"%>
 <%
 	session = request.getSession(false);
-	Builder builder = new Builder();
+	BuilderEmployee builder = new BuilderEmployee();
 	List<BuilderProject> project_list = null; 
 	int builder_id = 0;
 	if(session!=null)
 	{
 		if(session.getAttribute("ubname") != null)
 		{
-			builder  = (Builder)session.getAttribute("ubname");
-			builder_id = builder.getId();
+			builder  = (BuilderEmployee)session.getAttribute("ubname");
+			builder_id = builder.getBuilder().getId();
 		}
    }
 	int builder_id1 = 1;
