+++ src/main/webapp/builder/cancellation/list.jsp
@@ -16,14 +16,14 @@
 	List<CancellationList> cancellation_list = null;
 	List<ProjectData> projectDatas = null;
 	session = request.getSession(false);
-	Builder builder = new Builder();
+	BuilderEmployee builder = new BuilderEmployee();
 	List<City> cityList = new CityNamesImp().getCityNames();
 	int builder_uid = 0;
 	if(session!=null)
 	{
 		if(session.getAttribute("ubname") != null)
 		{
-			builder  = (Builder)session.getAttribute("ubname");
+			builder  = (BuilderEmployee)session.getAttribute("ubname");
 			builder_uid = builder.getId();
 		}
    	}
