+++ src/main/webapp/builder/changepassword.jsp
@@ -86,7 +86,7 @@
 <script type="text/javascript">
 function changePassword(){
 	if($("#npassword").val() == $("#cpassword").val()){
-		$.post('webapi/validate/builder/password/',{oldpassword: $("#opassword").val(), password: $("#npassword").val()}, function(data){
+		$.post('webapi/validatebuilder/builder/password/',{oldpassword: $("#opassword").val(), password: $("#npassword").val()}, function(data){
 			var success = data.status;
 			var status =parseInt(success);
 			if(status==1)
