+++ src/main/webapp/builder/employee/.gitignore
@@ -1,0 +1,3 @@
+/new.jsp
+/edit.jsp
+/old_add.jsp
