+++ src/main/webapp/builder/possession/.gitignore
@@ -1,0 +1,1 @@
+/list.jsp
