+++ src/main/webapp/builder/project/.gitignore
@@ -1,0 +1,2 @@
+/oldlist.jsp
+/list.jsp
