+++ src/main/webapp/builder/project/building/.gitignore
@@ -1,0 +1,1 @@
+/oldedit.jsp
