+++ src/main/webapp/builder/project/building/edit.jsp
@@ -26,13 +26,13 @@
 	int p_user_id = 0;
 	building_id = Integer.parseInt(request.getParameter("building_id"));
 	session = request.getSession(false);
-	Builder adminuserproject = new Builder();
+	BuilderEmployee adminuserproject = new BuilderEmployee();
 	if(session!=null)
 	{
 		if(session.getAttribute("ubname") != null)
 		{
-			adminuserproject  = (Builder)session.getAttribute("ubname");
-			p_user_id = adminuserproject.getId();
+			adminuserproject  = (BuilderEmployee)session.getAttribute("ubname");
+			p_user_id = adminuserproject.getBuilder().getId();
 		}
 	}
 	BuilderBuilding builderBuilding = null;
