+++ src/main/webapp/builder/project/building/floor/edit.jsp
@@ -20,13 +20,13 @@
 	int p_user_id = 0;
 	floor_id = Integer.parseInt(request.getParameter("floor_id"));
 	session = request.getSession(false);
-	Builder adminuserproject = new Builder();
+	BuilderEmployee adminuserproject = new BuilderEmployee();
 	if(session!=null)
 	{
 		if(session.getAttribute("ubname") != null)
 		{
-			adminuserproject  = (Builder)session.getAttribute("ubname");
-			p_user_id = adminuserproject.getId();
+			adminuserproject  = (BuilderEmployee)session.getAttribute("ubname");
+			p_user_id = adminuserproject.getBuilder().getId();
 		}
 	}
 	BuilderFloor builderFloor = null;
