+++ src/main/webapp/builder/project/building/floor/flat/edit.jsp
@@ -27,13 +27,13 @@
 	int floor_id = 0;
 	flat_id = Integer.parseInt(request.getParameter("flat_id"));
 	session = request.getSession(false);
-	Builder adminuserproject = new Builder();
+	BuilderEmployee adminuserproject = new BuilderEmployee();
 	if(session!=null)
 	{
 		if(session.getAttribute("ubname") != null)
 		{
-			adminuserproject  = (Builder)session.getAttribute("ubname");
-			p_user_id = adminuserproject.getId();
+			adminuserproject  = (BuilderEmployee)session.getAttribute("ubname");
+			p_user_id = adminuserproject.getBuilder().getId();
 		}
 	}
 	List<BuilderBuilding> buildings = null;
