+++ src/main/webapp/builder/project/building/floor/list.jsp
@@ -10,13 +10,13 @@
 	List<BuilderFloor> builderFloors = null;
 	List<ProjectData> projectDatas = null;
 	session = request.getSession(false);
-	Builder adminuserproject = new Builder();
+	BuilderEmployee adminuserproject = new BuilderEmployee();
 	if(session!=null)
 	{
 		if(session.getAttribute("ubname") != null)
 		{
-			adminuserproject  = (Builder)session.getAttribute("ubname");
-			p_user_id = adminuserproject.getId();
+			adminuserproject  = (BuilderEmployee)session.getAttribute("ubname");
+			p_user_id = adminuserproject.getBuilder().getId();
 		}
 	}
 	if (request.getParameterMap().containsKey("building_id")) {
