+++ src/main/webapp/builder/project/list.jsp
@@ -17,14 +17,14 @@
 	List<ProjectList> project_list = null;
 	ProjectImageGallery imageGaleries = null;
 	session = request.getSession(false);
-	Builder builder = new Builder();
+	BuilderEmployee builder = new BuilderEmployee();
 	int builder_uid = 0;
 	if(session!=null)
 	{
 		if(session.getAttribute("ubname") != null)
 		{
-			builder  = (Builder)session.getAttribute("ubname");
-			builder_uid = builder.getId();
+			builder  = (BuilderEmployee)session.getAttribute("ubname");
+			builder_uid = builder.getBuilder().getId();
 		}
 		if(builder_uid > 0){
 			project_list = new ProjectDAO().getBuilderActiveProjectsByBuilderId(builder_uid);
